substantive: remove Ordering section
In the context of this document, ordering is only used as part of a suggestion
about how authors may choose which version to reference, but I'm not even
convinced that this is a generally applicable suggestion. Instead of defining a
whole subsection and algorithm just for this one suggestion, I changed the
suggestion to be a bit broader.