Processors must validate directive definitions
Following discussion with @pcmanus and @martijnwalraven, we've decided that
compliant processors should expect directives to be defined with arguments
matching what the specification expects. To define a directive with an argument
whose type or default value differs from the specification is now an error that
should be detected, rather than something that could lead to surprising
processing (eg, a processor treating `@core__export` as equivalent to
`@core__export(export: true)` even if a specific schema defines the
directive as `directive @core__export(export: Boolean! = false)`.

Specifically, what we validate is:
- All expected arguments are provided with the correct types and default values
- No additional arguments are provided
- The `repeated` flag is as expected
- The list of locations is as expected

We do not validate:
- Order of arguments or locations
- Description string on the directive or any arguments
- Directives applied to the argument definitions

This should be sufficient to ensure that, once the definitions have been
validated, later processing code can assume that directive arguments have the
expected types and the semantics of omitting out an argument is well-defined.

We remove from the spec the concept of the extensibility of directives (and
other types). We still think this is a good idea, but we'd like to wait until we
have a use case for it. It makes defining this validation simpler if it doesn't
have to include sometimes allowing additional arguments.

Include a top-level section with both directive definitions that can be easily
copied into core schemas. Update most examples to refer to this section instead
of trying to define the directives.