Remove customizability of export from spec
There's some disagreement on the team as to whether this belongs in Core Schema
itself or something separate. The only planned short-term use case for these
features is the "inaccessible fields" proposal and there is disagreement as to
whether `@core__export` is actually appropriate for that use case.

It is very likely we will add something like this back sometime but let's keep
it out of v0.1 for now.

It turns out that it now is easy for us to entirely remove the term "export"
from the spec and stick with the newer "API" term.

It also means that actually declaring the directive in every example is
non-burdensome so we do that.

Also remove references to extensions of foreign types since we don't have great
examples/use cases for that right now.