Clarify that trailing slashes in the feature URL should be ignored
Core schema processors already need to parse the URL and ignore the query string
and fragment; ignoring a trailing slash as well seems reasonable, and it may
make it easier to serve the human-readable spec at the feature URL if the
feature URL may contain a trailing slash, as many web hosting setups prefer
trailing slashes on URLs.

Note that the identity calculated from the feature URL by dropping the version
tag still never has a trailing slash.

Clarify some other text (eg, replace "prefix" with "name" in one place, because
"name" is what we settled on for calling that part of the URL).