Remove core-schema link/header ?
This renders as a link in my setup although I cannot click it. Maybe there's a better fix, I'm not sure what the initial intent was, looks like this was designed to do something else but I can't figure out why. 
Feel free to discard this PR if I'm missing something.